requirements.txt
PdfReader
pdfminer.six
nltk
pytesseract
pillow
pdf2image